# api/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/account/hello": {
            "get": {
                "description": "向账号服务发送问候请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "账号服务问候接口",
                "parameters": [
                    {
                        "description": "问候参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.HelloRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.HelloResponse"
                        }
                    }
                }
            }
        },
        "/account/login": {
            "post": {
                "description": "用户登录账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "账号登录接口",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.LoginResponse"
                        }
                    }
                }
            }
        },
        "/account/register": {
            "post": {
                "description": "用户注册账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "账号注册接口",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_account_ginhandler.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/s3/PutKey": {
            "post": {
                "description": "上传文件内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "S3"
                ],
                "summary": "上传文件到 S3",
                "parameters": [
                    {
                        "description": "上传参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ginhandler.PutKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginhandler.PutKeyResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api_account_ginhandler.HelloRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "名字",
                    "type": "string",
                    "example": "John"
                }
            }
        },
        "api_account_ginhandler.HelloResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "Hello, John"
                }
            }
        },
        "api_account_ginhandler.LoginRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "账号",
                    "type": "string",
                    "example": "john_doe"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "securepassword"
                }
            }
        },
        "api_account_ginhandler.LoginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "Login successful"
                },
                "token": {
                    "description": "认证令牌",
                    "type": "string",
                    "example": "some-token"
                }
            }
        },
        "api_account_ginhandler.RegisterRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "用户名",
                    "type": "string",
                    "example": "john"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "example": "john@example.com"
                },
                "name": {
                    "description": "姓名",
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "securepassword"
                }
            }
        },
        "api_account_ginhandler.RegisterResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer",
                    "example": 201
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "Registration successful"
                }
            }
        },
        "ginhandler.PutKeyRequest": {
            "type": "object",
            "properties": {
                "bucket": {
                    "description": "桶名称",
                    "type": "string",
                    "example": "my-bucket"
                },
                "data": {
                    "description": "文件内容",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "key": {
                    "description": "文件路径",
                    "type": "string",
                    "example": "images/logo.png"
                }
            }
        },
        "ginhandler.PutKeyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "PutKey successful"
                }
            }
        }
    }
}